pyproject.toml
[project]
name = "evlampiy-notes-tgbot"
version = "0.2.2"
description = "Telegram bot (Evlampiy) for managing notes"
readme = "README.md"
authors = ["Yast <yastcher@gmail.com>"]
requires-python = "~=3.12"
dependencies = [
    "python-telegram-bot~=21.11.1",
    "httpx[http2]>=0.28.1",
    "python-dotenv~=1.0.1",
    "pydantic-settings~=2.7.1",
    "wit~=6.0.1",
    "pydub~=0.25.1",
    "openai~=1.59.8",
    "beanie==1.29.0",
    "motor==3.7.1",
]

[dependency-groups]
dev = [
    "black~=24.4.2",
    "pytest~=8.4.2",
    "pytest-asyncio~=0.23.8",
    "pytest-mock==3.15.1",
    "pytest-cov~=7.0.0",
]

[tool.ruff]
line-length = 120
exclude = [
    '.github',
]

[tool.ruff.lint]
select = ["E", "F", "N", "I"]
ignore = [
    "E203", "E501", "E731",
    "N802", "N803", "N805", "N806",
]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-local-folder = [
    'src',
    'tests',
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
# skip-magic-trailing-comma = true

[tool.pytest.ini_options]
addopts = [
    "-p", "pytest_cov",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=json:coverage.json",
    "--cov-branch"
]
testpaths = ["tests"]
